requirements.txt
numpy==1.22.2
pandas==1.4.1
Pillow==9.0.1
scikit_learn==1.0.2
streamlit==1.6.0
